Added callback on onPanComplete.
Calls back when pan() is called, or when onMouseUp fires.
The callback is called with the following signature:
function callback([x, y], transform).

To make the callbacks more powerful, onZoom was changed from
onZoom(scale) to onZoom(scale, viewportTransform). This way
the entire transform can be retrieved from either callback.